refactor test functions to util file
