+++ test/activity/description/d2l-activity-description-editor.test.js
@@ -1,5 +1,6 @@
 import '../../../components/activity/description/d2l-activity-description-editor.js';
-import { assert, aTimeout, elementUpdated, expect, fixture, html } from '@open-wc/testing';
+import { assert,  elementUpdated, expect } from '@open-wc/testing';
+import { createComponentAndWait, delayAndAwaitForElement, fireTextareaInputEvent } from '../test-util.js';
 import { learningPathExisting, learningPathNew, learningPathUpdated } from '../../data.js';
 import { clearStore } from '@brightspace-hmc/foundation-engine/state/HypermediaState.js';
 import { mockLink } from '../../fetchMocks.js';
@@ -9,33 +10,8 @@ import sinon from 'sinon/pkg/sinon-esm.js';
 // use the learningPathUpdated description as the text when updating textareas
 const updatedDescriptionText = learningPathUpdated.properties.description;
 
-async function _createComponentAndWait(path)
-{
-	const element = await fixture(html`<d2l-activity-description-editor href="${path}" token="test-token"></d2l-activity-description-editor>`);
-	await element.updateComplete;
-	await element._state.allFetchesComplete();
-	await element.updateComplete;
-	return element;
-}
-
-// using a delay to wait for fetchMock to be called
-// would prefer a better way
-async function _delayAndAwaitForElement(element, delayMs) {
-	return aTimeout(delayMs).then(async() => {
-		await elementUpdated(element);
-	});
-}
-
-// Handle updating the component textarea with a new value and await the update
-async function fireTextareaInputEvent(element) {
-	const textarea = element.shadowRoot.querySelector('label textarea');
-	textarea.value = updatedDescriptionText;
-	const inputEvent = new Event('input');
-	textarea.dispatchEvent(inputEvent);
-
-	// A render happens after events, we needed this extra await
-	// wait for element to be updated from the input event
-	await elementUpdated(element);
+async function _createDescriptionEditor(path) {
+	return await createComponentAndWait(path);
 }
 
 describe('d2l-activity-description-editor', () => {
@@ -58,7 +34,7 @@ describe('d2l-activity-description-editor', () => {
 		});
 
 		it.skip('should initialize using defined path and expected values', async() => {
-			const element = await _createComponentAndWait('/learning-path/new');
+			const element = await _createDescriptionEditor('/learning-path/new');
 
 			// paths should be followed
 			assert.isTrue(mockLink.called('path:/learning-path/new'), '/learing-path/new was not called');
@@ -85,7 +61,7 @@ describe('d2l-activity-description-editor', () => {
 			let element;
 			beforeEach(async() => {
 				clearStore();
-				element = await _createComponentAndWait('/learning-path/existing');
+				element = await _createDescriptionEditor('/learning-path/existing');
 				assert.equal(element.description, learningPathExisting.properties.description, 'description should match response');
 			});
 
@@ -102,7 +78,7 @@ describe('d2l-activity-description-editor', () => {
 
 			it('updating should commit state', async() => {
 				const spy = sinon.spy(element.updateDescription);
-				await fireTextareaInputEvent(element);
+				await fireTextareaInputEvent(element, updatedDescriptionText);
 
 				assert.equal(element.description, updatedDescriptionText, 'description was updated to match');
 				assert.isTrue(spy.commit.called, 'Commit should be called when input event is triggered');
@@ -112,10 +88,10 @@ describe('d2l-activity-description-editor', () => {
 			*	Still need to figure out how to wait for element to be updated after push / reset
 			*/
 			it('pushing state should save description', async() => {
-				await fireTextareaInputEvent(element);
+				await fireTextareaInputEvent(element, updatedDescriptionText);
 
 				element._state.push();
-				await _delayAndAwaitForElement(element, 100);
+				await delayAndAwaitForElement(element, 100);
 
 				assert.isTrue(mockLink.called('path:/description/update'), 'Updated path was not called');
 				assert.equal(element.description, updatedDescriptionText, 'description should be updated after a push');
@@ -126,7 +102,7 @@ describe('d2l-activity-description-editor', () => {
 			});
 
 			it('reset state should revert description and textarea', async() => {
-				await fireTextareaInputEvent(element);
+				await fireTextareaInputEvent(element, updatedDescriptionText);
 
 				assert.equal(element.description, updatedDescriptionText, 'description should be updated');
